Don't call API from views
I'm new to Flux, but from what I understand (particularly from the
helpful diagram on the index page of this app), it is not the Flux Way
to call APIs directly from views. Instead, the views should call an
Action Creator, which itself can call an API. And then when the API
response comes back, it also goes through an Action Creator.

I've reworked this app to follow that principle. It also now remembers
the todos between page refreshes.